support small-bucket brigades, see #2
With `EnableMMAP On`, the brigade of a 75k file looks roughly like this:

75k -> EOS

When `EnableMMAP Off` is set in Apache, the brigade is split into a lot
of ≈8000b buckets:

8k -> 8k -> 8k -> 8k -> 8k -> 8k -> 8k -> 8k -> 8k -> 3k -> EOS

All chunks are smaller than brotli's `block_space` so we never call
`brotli_compress()`, but the buckets payload is still copied into
brotlis ring buffer. When EOS is reached, we find that it has grown too
big.

Now, the brotli context as a `ring_size` variable that tracks the size
of the ring buffer. Whenever we are about to hit the `block_space` size,
we call `brotli_compress()` and reset the `ring_size`.

Thus, we avoid calling brotli with not completely filled buffers,
because that could result in a less compressed output.